+++ packages/sdk/browser/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [4.3.1](https://github.com/launchdarkly/js-core/compare/js-client-sdk-v4.3.0...js-client-sdk-v4.3.1) (2026-03-04)
+
+
+### Bug Fixes
+
+* **js-client-sdk:** switching context does not update streaming connection ([#1153](https://github.com/launchdarkly/js-core/issues/1153)) ([b894ac2](https://github.com/launchdarkly/js-core/commit/b894ac29b3d054f88d4e0c3ee24fcb59ea53cca9))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-client-sdk-common bumped from 1.20.0 to 1.21.0
+
 ## [4.3.0](https://github.com/launchdarkly/js-core/compare/js-client-sdk-v4.2.1...js-client-sdk-v4.3.0) (2026-02-25)